Code Review Comments
Great addition of the count down! One suggestion would be to use "let" instead of "var" since ECMA 6 is now using "let" to prevent errors and bugs caused by scoping. Also perhaps a little CSS to the countdown would make it stand out more since it's such nice touch to the website, it would be great to have it stand out more!