Peer review comments
Nice job on the CSS! The animated bars are really cool and so is the tilting of the page when the mouse hovers over the sidebar! I think it would be even better if you put CSS files separately for different pages so you can have nicely organized CSS files that's clear on what css stylesheet is used for which HTML page.